adds axios for spotper api/changes states to redux
