+++ src/services/api.js
@@ -6,10 +6,10 @@ const spotifyApi = axios.create({
     baseURL: 'https://api.spotify.com/v1'
 })
 
-//const spotperApi = axios.create({
-//    baseURL: 'http://localhost:8080/rulesview/api'
-//})
+const spotperApi = axios.create({
+    baseURL: 'http://localhost:8080/api/v1'
+})
 
 
 
-export { spotifyApi }
+export { spotifyApi, spotperApi }
